% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributed_modelling.R
\name{trainDistributedModel}
\alias{trainDistributedModel}
\title{Train Distributed Model}
\usage{
trainDistributedModel(regis_dir, silent = FALSE, epochs_at_once = 1L)
}
\arguments{
\item{regis_dir}{[\code{character(1)}]\cr
Direction for the output files.}

\item{silent}{[\code{logical(1)}]\cr}

\item{epochs_at_once}{[\code{integer(1)}]\cr}
}
\description{
This function conducts a specific number of epochs on the local machine. Therefore, the function reads in the available
datasets and runs a fixed number of Gradient Descent steps.
}
